created monitor.py
script to monitor tcp and http ports
Line 1 is a requisite for running a python script directly (i.e. without invoking python first).
Lines 3-5 load the modules for HTTP(S) connections, TCP connections, and command line arguments, respectively.
Lines 7-15 establish a TCP connection to an arbitrary hostname:port, and report if the attempt was successful.
Lines 17-22 establish makes an HTTP(S) request to a server, and reports if the attempt was successful.
Lines 24-28 parse the arguments passed to the script, and determine which test was requested.
Lines 30-34 are invoked when the script is initially called on the command line. First the number of arguments is checked, then the actual server test is run. If the test fails, it prints an error message.

monitor.py tcp xxx.xxx.xxx.xxx:22
monitor.py http <a href="http://xxx.xxx.xxx/">http://xxx.xxx.xxx/</a>