Create monitor_alerts.py
Line 3 imports the system() call from the os module. This will allow the monitor to execute other programs.
Line 7 imports asctime from the time module. This will be put in the email body so that uptime can be accurately gagued.
Line 32-35 build the email message, and invoke the mail program.
Line 38 allows an additional argument: an email address.
Line 41 changes the print command to an email command, instead of writing an error message to the screen, the error will now be sent to an email account of someone who cares.